requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cryptography==38.0.3
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.3
django-appconf==1.0.5
django-autoslug==1.9.8
django-bootstrap-modal-forms==2.2.0
django-cities-light==3.9.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-filter==22.1
django-fontawesome-5==1.0.18
django-haystack==3.2.1
django-otp==1.1.4
django-payments==1.0.0
django-payments-razorpay==1.0
django-paypal==2.0
django-phonenumber-field==7.0.0
django-redis==5.2.0
django-selectable-select2==0.4.0
django-staticfiles-select2==3.4.5
django-summernote==0.8.20.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
elastic-transport==8.4.0
elasticsearch==5.5.3
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
phonenumbers==8.13.0
phonenumberslite==8.13.0
Pillow==9.3.0
progressbar2==4.2.0
prompt-toolkit==3.0.32
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyotp==2.7.0
pyparsing==3.0.9
python-decouple==3.6
python-utils==3.4.5
python3-openid==3.2.0
pytz==2022.6
razorpay==1.3.0
redis==4.3.4
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
sendotp==0.0.4
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Whoosh==2.7.4
wrapt==1.14.1
